Docker-ify bandcamp scraper for cloud run
