+++ functions/bandcamp-user-data/package.json
@@ -5,7 +5,7 @@
   "main": "dist/index.js",
   "license": "MIT",
   "scripts": {
-    "start": "nodemon -e js,json,ts -w src -x \"tsc && functions-framework --target=routeHandler --port=8081 dist/index.js\"",
+    "start": "nodemon -e js,json,ts -w src -x \"ts-node src/index.ts\"",
     "build": "tsc -p .",
     "test": "jest",
     "gcp-build": "tsc -p .",
@@ -14,7 +14,8 @@
   "dependencies": {
     "@google-cloud/functions-framework": "^1.7.1",
     "express": "^4.17.1",
-    "playwright": "^1.7.1"
+    "playwright": "^1.7.1",
+    "ts-node": "^9.1.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.9",
