+++ functions/bandcamp-user-data/src/__tests__/pullBandcampData.spec.ts
@@ -1,11 +1,11 @@
-import { pullBandcampData, RawBandcampUserData } from '..';
+import { pullBandcampData, RawBandcampUserData } from '../utils';
 
 /** The main goal with tests here is to just make sure that the correct data is still being returned */
 describe('pullBandcampData', () => {
   it('still works with an existing account', async () => {
     const expectedData: Partial<RawBandcampUserData> = {
-      username: process.env.BC_TEST_USER,
-      fan_id: parseInt(process.env.BC_TEST_ID as string),
+      username: process.env.BC_TEST_USER || 'papa_lasagna',
+      fan_id: parseInt(process.env.BC_TEST_ID || '4335879'),
     };
     const received = await pullBandcampData(expectedData.username as string);
